+++ CHANGES.md
@@ -78,6 +78,7 @@ Changes in 10.0
 - Compatibility with add-ons version < 10.0, gather CONFIG (uppercase table name) values too in Configuration.php
 - Fix MySQL error Table is specified twice, both as a target for 'INSERT' and as a separate source for data in CopySchool.php & Rollover.php
 - Fix MySQL syntax error: no table alias in DELETE in Rollover.php
+- Fix MySQL syntax error: no FROM allowed inside UPDATE, use subquery or multi-table syntax in Rollover.php
 
 Changes in 9.3.1
 ----------------